// apis/wutopia/types.ts
export type WutopiaCartoonData = {
  acrossPic: string;
  acrossPicUrl: string;
  acrossPicUrlWebp: string;
  author: string;
  autoShelvesState: number;
  browerChapterId: number;
  browseCount: number;
  cartoonPic: string;
  cartoonPicUrl: string;
  cartoonPicUrlWebp: string;
  cartoonTypes: WutopiaCartoonTypeData[];
  channel: number;
  channelStr: string;
  chapterCount: number;
  chapterIndex: number;
  collectCount: number;
  content: string;
  copyrightPartlyId: number;
  cpChannelId: number;
  cpComicId: string;
  cpUpdateTime: string;
  createTime: number;
  createTimeStr: string;
  dayNumber: number;
  first: {
    chapterIndex: number;
    id: number;
    modifyTimeStr: string;
  };
  frequency: number;
  id: number;
  isAppHide: number;
  isCpChannel: number;
  isFinish: number;
  isFinishStr: string;
  isHide: number;
  isShowEighteenTag: number;
  isShowSubject: boolean;
  lastDayMoney: number;
  maxChapterIndex: number;
  modifyState: number;
  modifyStateStr: string;
  modifyTime: number;
  modifyTimeStr: string;
  name: string;
  payState: number;
  payStateStr: string;
  pic: string;
  picUrl: string;
  picUrlWebp: string;
  planShelvesTime: number;
  praiseCount: number;
  price: number;
  probationChapterCount: number;
  probationChapterIds: string;
  rating: string;
  recoBrowseClickCount: number;
  recoCollectClickCount: number;
  recoCollectDeleteCount: number;
  recoIndex: number;
  shamBrowseCount: number;
  shamBrowseCountStr: string;
  shamCollectCount: number;
  shamPraiseCount: number;
  shelvesState: number;
  shelvesTime: number;
  shelvesType: number;
  subscribeCount: number;
  typeNames: string;
  updateSchedule: string;
  updateScheduleStr: string;
};

export type WutopiaCartoonTypeData = {
  color: string;
  id: number;
  name: string;
};

export type WutopiaChapterData = {
  autoShelvesState: number;
  browseCount: number;
  cartoonCollectionId: number;
  chapterIndex: number;
  coverPic: string;
  coverPicUrl: string;
  cpChapterId: string;
  cpComicId: string;
  cpUpdateTime: string;
  createTime: number;
  id: number;
  isCpChannel: number;
  isPayed: boolean;
  isPraise: boolean;
  isPurchase: boolean;
  isRead: boolean;
  modifyTime: number;
  modifyTimeStr: string;
  name: string;
  nextChapter?: {
    cartoonCollectionId: number;
    chapterIndex: number;
    id: number;
    modifyTimeStr: string;
  };
  picList?: WutopiaChapterPictureData[];
  planShelvesTime: number;
  praiseCount: number;
  remark: string;
  shamBrowseCount: number;
  shamPraiseCount: number;
  shelvesDateStr: string;
  shelvesState: number;
  shelvesTime: number;
  shelvesType: number;
  subscribeCount: number;
};

export type WutopiaChapterPictureData = {
  chapterId: number;
  cpChapterId: string;
  height: number;
  id: number;
  pic: string;
  picUrl: string;
  picUrlWebp: string;
  picWaterUrl: string;
  width: number;
};